// model/userData.ts
export default interface UserData {
    id?: number;
    nome?: string;
    sobrenome?: string;
    email?: string;
    imagem?: string;
    countrycode?: string;
    telemovel?: string;
    nacionalidade?: number;
    zipcode?: string,
    data_nascimento?: string,
    identificacao?: string,
    password?: string,
    verified?: string,
    notif?: string,
    sms?: string,
    registo?: string,
    saldo?: string,
}